refactor: move Banner directory out to components
